Now with Charts!
Added the option to show you a chart of the progress of each player, round to round, at the end of the game. Also, minor adjustments, such as a few stylistic choices (the modal windows now resize without overflowing the screen) and the addition of a button directing towards the repository.